Finished: - Delete Functionality added for Workouts - Undid code cleanup
TO DO:
- do more research on whether it's okay to throw exceptions in the service classes
- add update functionality for workouts
- make table with exercise names
- add dropdown for exercise selection on addExercise page
- Add warning page prior to deletion
- Make app stateless and disable csrf

+++ src/main/java/org/hsha/hsha/controllers/ExerciseController.java
@@ -38,8 +38,15 @@ public class ExerciseController {
 
 
         // Validate the user and workout
-        Optional<User> user = userService.verifyUser(userId);
-        Optional<Workout> userWorkout = workoutService.verifyWorkout(workoutId, user);
+        Optional<User> user = userService.retrieveUserById(userId);
+        if(user.isEmpty()) {
+            throw new ServerException("User: " + userId + " not found");
+        }
+        Optional<Workout> userWorkout = workoutService.retrieveWorkoutById(workoutId);
+        if(userWorkout.isEmpty() || !(userWorkout.get().getUser().getId().equals( user.get().getId()))) {
+            throw new ServerException("Workout: " + workoutId + " not found");
+        }
+
 
 
         ExerciseDto exerciseDto = new ExerciseDto();
@@ -53,8 +60,15 @@ public class ExerciseController {
     public String addExerciseToWorkout(@PathVariable int userId, @PathVariable int workoutId,
                                        @ModelAttribute ExerciseDto exerciseDto, BindingResult result) throws ServerException {
         // Validate the user and workout
-        Optional<User> user = userService.verifyUser(userId);
-        Optional<Workout> userWorkout = workoutService.verifyWorkout(workoutId, user);
+        Optional<User> user = userService.retrieveUserById(userId);
+        if(user.isEmpty()) {
+            throw new ServerException("User: " + userId + " not found");
+        }
+        Optional<Workout> userWorkout = workoutService.retrieveWorkoutById(workoutId);
+        if(userWorkout.isEmpty() || !(userWorkout.get().getUser().getId().equals( user.get().getId()))) {
+            throw new ServerException("Workout: " + workoutId + " not found");
+        }
+
 
         try {
             Exercise newExercise = new Exercise();
@@ -74,9 +88,18 @@ public class ExerciseController {
     public String showUpdateExercisePage(@PathVariable int userId,@PathVariable int workoutId, @PathVariable int exerciseId, Model model)
             throws ServerException {
         // Validate the user, workout, and exercises
-        Optional<User> user = userService.verifyUser(userId);
-        Optional<Workout> userWorkout = workoutService.verifyWorkout(workoutId, user);
-        Optional<Exercise> workoutExercise = exerciseService.verifyExercise(exerciseId, userWorkout);
+        Optional<User> user = userService.retrieveUserById(userId);
+        if(user.isEmpty()) {
+            throw new ServerException("User: " + userId + " not found");
+        }
+        Optional<Workout> userWorkout = workoutService.retrieveWorkoutById(workoutId);
+        if(userWorkout.isEmpty() || !(userWorkout.get().getUser().getId().equals( user.get().getId()))) {
+            throw new ServerException("Workout: " + workoutId + " not found");
+        }
+        Optional<Exercise> workoutExercise = exerciseService.retrieveExerciseById(exerciseId);
+        if(workoutExercise.isEmpty() || !(workoutExercise.get().getWorkout().getId().equals(userWorkout.get().getId()))) {
+            throw new ServerException("Exercise: " + exerciseId + " not found");
+        }
 
         ExerciseDto exerciseDto = new ExerciseDto();
         model.addAttribute("user", user);
@@ -92,9 +115,18 @@ public class ExerciseController {
                                         BindingResult result)
             throws ServerException {
         // Validate the user, workout, and exercises
-        Optional<User> user = userService.verifyUser(userId);
-        Optional<Workout> userWorkout = workoutService.verifyWorkout(workoutId, user);
-        Optional<Exercise> workoutExercise = exerciseService.verifyExercise(exerciseId, userWorkout);
+        Optional<User> user = userService.retrieveUserById(userId);
+        if(user.isEmpty()) {
+            throw new ServerException("User: " + userId + " not found");
+        }
+        Optional<Workout> userWorkout = workoutService.retrieveWorkoutById(workoutId);
+        if(userWorkout.isEmpty() || !(userWorkout.get().getUser().getId().equals( user.get().getId()))) {
+            throw new ServerException("Workout: " + workoutId + " not found");
+        }
+        Optional<Exercise> workoutExercise = exerciseService.retrieveExerciseById(exerciseId);
+        if(workoutExercise.isEmpty() || !(workoutExercise.get().getWorkout().getId().equals(userWorkout.get().getId()))) {
+            throw new ServerException("Exercise: " + exerciseId + " not found");
+        }
 
         try {
             workoutExercise.get().setWorkout(userWorkout.get());
@@ -113,9 +145,18 @@ public class ExerciseController {
     public String showDeleteExercisePage(@PathVariable int userId,@PathVariable int workoutId,
                                          @PathVariable int exerciseId, Model model) throws ServerException {
         // Validate the user, workout, and exercises
-        Optional<User> user = userService.verifyUser(userId);
-        Optional<Workout> userWorkout = workoutService.verifyWorkout(workoutId, user);
-        Optional<Exercise> workoutExercise = exerciseService.verifyExercise(exerciseId, userWorkout);
+        Optional<User> user = userService.retrieveUserById(userId);
+        if(user.isEmpty()) {
+            throw new ServerException("User: " + userId + " not found");
+        }
+        Optional<Workout> userWorkout = workoutService.retrieveWorkoutById(workoutId);
+        if(userWorkout.isEmpty() || !(userWorkout.get().getUser().getId().equals( user.get().getId()))) {
+            throw new ServerException("Workout: " + workoutId + " not found");
+        }
+        Optional<Exercise> workoutExercise = exerciseService.retrieveExerciseById(exerciseId);
+        if(workoutExercise.isEmpty() || !(workoutExercise.get().getWorkout().getId().equals(userWorkout.get().getId()))) {
+            throw new ServerException("Exercise: " + exerciseId + " not found");
+        }
         List<ExSet> exerciseExSets = exSetService.retrieveAllExSetsByWorkoutId(workoutId);
         ThIterationCounter counter = new ThIterationCounter();
 
@@ -133,9 +174,18 @@ public class ExerciseController {
     @RequestMapping("users/{userId}/workouts/{workoutId}/exercises/{exerciseId}/delete/request")
     public String deleteExercise(@PathVariable int userId,@PathVariable int workoutId,
                                  @PathVariable int exerciseId) throws ServerException {
-        Optional<User> user = userService.verifyUser(userId);
-        Optional<Workout> userWorkout = workoutService.verifyWorkout(workoutId, user);
-        Optional<Exercise> workoutExercise = exerciseService.verifyExercise(exerciseId, userWorkout);
+        Optional<User> user = userService.retrieveUserById(userId);
+        if(user.isEmpty()) {
+            throw new ServerException("User: " + userId + " not found");
+        }
+        Optional<Workout> userWorkout = workoutService.retrieveWorkoutById(workoutId);
+        if(userWorkout.isEmpty() || !(userWorkout.get().getUser().getId().equals( user.get().getId()))) {
+            throw new ServerException("Workout: " + workoutId + " not found");
+        }
+        Optional<Exercise> workoutExercise = exerciseService.retrieveExerciseById(exerciseId);
+        if(workoutExercise.isEmpty() || !(workoutExercise.get().getWorkout().getId().equals(userWorkout.get().getId()))) {
+            throw new ServerException("Exercise: " + exerciseId + " not found");
+        }
 
         exerciseService.deleteExerciseById(workoutExercise.get().getId());
